Rearrange -> API (#111)
Co-authored-by: SFYLL <sd@seismic.systems>

restructures the enclave repo code. main feature improvement is concurrency improvements. A central object owns the attestation agent and the verifier with a mutex so that the tpm and other resources that must be accessed in a single threaded way do not get overwhelmed

this PR is heavy on the code that restructures the verifier in particular